+++ e2e/react-start/split-base-and-basepath/vite.config.ts
@@ -1,17 +1,14 @@
 import { defineConfig } from 'vite'
-import tsConfigPaths from 'vite-tsconfig-paths'
 import { tanstackStart } from '@tanstack/react-start/plugin/vite'
 import viteReact from '@vitejs/plugin-react'
 
 export default defineConfig({
+  resolve: { tsconfigPaths: true },
   base: '/_ui/',
   server: {
     port: 3000,
   },
   plugins: [
-    tsConfigPaths({
-      projects: ['./tsconfig.json'],
-    }),
     tanstackStart({
       router: {
         basepath: '/',