+++ e2e/solid-start/basic-cloudflare/vite.config.ts
@@ -1,16 +1,13 @@
 import { defineConfig } from 'vite'
-import tsConfigPaths from 'vite-tsconfig-paths'
 import { cloudflare } from '@cloudflare/vite-plugin'
 import { tanstackStart } from '@tanstack/solid-start/plugin/vite'
 import viteSolid from 'vite-plugin-solid'
 import tailwindcss from '@tailwindcss/vite'
 
 export default defineConfig({
+  resolve: { tsconfigPaths: true },
   plugins: [
     tailwindcss(),
-    tsConfigPaths({
-      projects: ['./tsconfig.json'],
-    }),
     cloudflare({ viteEnvironment: { name: 'ssr' }, inspectorPort: false }),
     tanstackStart({
       prerender: {