+++ packages/vue-router/eslint.config.ts
@@ -3,6 +3,9 @@ import rootConfig from '../../eslint.config.js'
 export default [
   ...rootConfig,
   {
+    ignores: ['bin/**'],
+  },
+  {
     files: ['src/**/*.{ts,tsx}', 'tests/**/*.{ts,tsx}'],
     plugins: {},
     rules: {