+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @tanstack/router-cli
 
+## 1.166.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.15
+
+## 1.166.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.14
+
+## 1.166.14
+
+### Patch Changes
+
+- fix(router-cli): pass process.argv to yargs to fix silent CLI failure ([#6981](https://github.com/TanStack/router/pull/6981))
+
+## 1.166.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.13
+
+## 1.166.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.166.12
+
 ## 1.166.11
 
 ### Patch Changes