+++ packages/solid-router-ssr-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @tanstack/solid-router-ssr-query
 
+## 1.166.10
+
+### Patch Changes
+
+- Updated dependencies [[`0545239`](https://github.com/TanStack/router/commit/054523900b2ee19308e5a88417dadfc6923afe30)]:
+  - @tanstack/router-ssr-query-core@1.167.0
+
 ## 1.166.9
 
 ### Patch Changes